consumer/requirements.txt
fastapi
uvicorn
psycopg2-binary
kafka-python
requests
streamlit
plotly
Faker
streamlit-autorefresh
